Renamed RxLifecycleObserver to AndroidLifecycle
and changed constructor to factory method. This makes it better match
the navi provider.
