+++ rxlifecycle-android-lifecycle-kotlin/src/main/java/com/trello/rxlifecycle2/android/lifecycle/kotlin/rxlifecycle.kt
@@ -16,35 +16,35 @@ package com.trello.rxlifecycle2.android.lifecycle.kotlin
 
 import android.arch.lifecycle.Lifecycle
 import android.arch.lifecycle.LifecycleOwner
-import com.trello.lifecycle2.android.lifecycle.RxLifecycleObserver
+import com.trello.lifecycle2.android.lifecycle.AndroidLifecycle
 import io.reactivex.*
 
 fun <T> Observable<T>.bindToLifecycle(owner: LifecycleOwner): Observable<T>
-        = this.compose(RxLifecycleObserver(owner).bindToLifecycle())
+        = this.compose(AndroidLifecycle.createLifecycleProvider(owner).bindToLifecycle())
 
 fun <T> Observable<T>.bindUntilEvent(owner: LifecycleOwner, event: Lifecycle.Event): Observable<T>
-        = this.compose(RxLifecycleObserver(owner).bindUntilEvent(event))
+        = this.compose(AndroidLifecycle.createLifecycleProvider(owner).bindUntilEvent(event))
 
 fun <T> Flowable<T>.bindToLifecycle(owner: LifecycleOwner): Flowable<T>
-        = this.compose(RxLifecycleObserver(owner).bindToLifecycle())
+        = this.compose(AndroidLifecycle.createLifecycleProvider(owner).bindToLifecycle())
 
 fun <T> Flowable<T>.bindUntilEvent(owner: LifecycleOwner, event: Lifecycle.Event): Flowable<T>
-        = this.compose(RxLifecycleObserver(owner).bindUntilEvent(event))
+        = this.compose(AndroidLifecycle.createLifecycleProvider(owner).bindUntilEvent(event))
 
 fun <T> Single<T>.bindToLifecycle(owner: LifecycleOwner): Single<T>
-        = this.compose(RxLifecycleObserver(owner).bindToLifecycle())
+        = this.compose(AndroidLifecycle.createLifecycleProvider(owner).bindToLifecycle())
 
 fun <T> Single<T>.bindUntilEvent(owner: LifecycleOwner, event: Lifecycle.Event): Single<T>
-        = this.compose(RxLifecycleObserver(owner).bindUntilEvent(event))
+        = this.compose(AndroidLifecycle.createLifecycleProvider(owner).bindUntilEvent(event))
 
 fun <T> Maybe<T>.bindToLifecycle(owner: LifecycleOwner): Maybe<T>
-        = this.compose(RxLifecycleObserver(owner).bindToLifecycle())
+        = this.compose(AndroidLifecycle.createLifecycleProvider(owner).bindToLifecycle())
 
 fun <T> Maybe<T>.bindUntilEvent(owner: LifecycleOwner, event: Lifecycle.Event): Maybe<T>
-        = this.compose(RxLifecycleObserver(owner).bindUntilEvent(event))
+        = this.compose(AndroidLifecycle.createLifecycleProvider(owner).bindUntilEvent(event))
 
 fun <T> Completable.bindToLifecycle(owner: LifecycleOwner): Completable
-        = this.compose(RxLifecycleObserver(owner).bindToLifecycle<Completable>())
+        = this.compose(AndroidLifecycle.createLifecycleProvider(owner).bindToLifecycle<Completable>())
 
 fun <T> Completable.bindUntilEvent(owner: LifecycleOwner, event: Lifecycle.Event): Completable
-        = this.compose(RxLifecycleObserver(owner).bindUntilEvent<Completable>(event))
+        = this.compose(AndroidLifecycle.createLifecycleProvider(owner).bindUntilEvent<Completable>(event))
